requirements.txt
requests
pandas
matplotlib
rich
halo
pyinputplus
python-dotenv
ollama
